Turn off collect_ids when validation is off.
This commit turns off the collect_ids option of lxml when validation is
not needed.

The fact that xml:id values must be unique in a document is a validation
constraint, not a well-formedness constraint. However, even when
validation is not needed, lxml will raise an error on duplicate xml:ids
because lxml's collect_ids option is turned on by default. Effectively,
this means that a well-formed document will fail to parse even if
validation is not requested if it happens to have duplicate xml:id
values.