add mail-account.xsd as initial_schema_data
Introduce an initial_schema_data directory containing all the schema we
want to provide with eulxml. As of now, this schema only contains
``mail-account.xsd``, which was externally hosted until the link broke.

When generating the catalog, copy all the files in
``initial_schema_data`` in the catalog directory, and then download all
external schemas.

ref #40